# Support comparing BenchmarkDotNet results