--- a/src/Microsoft.Crank.Controller/ExecutionResult.cs
+++ b/src/Microsoft.Crank.Controller/ExecutionResult.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using Microsoft.Crank.Models;
 
 namespace Microsoft.Crank.Controller
@@ -11,5 +12,7 @@ namespace Microsoft.Crank.Controller
         public int ReturnCode { get; set; }
 
         public JobResults JobResults { get; set; } = new JobResults();
+
+        public Benchmark[] Benchmarks { get; set; } = Array.Empty<Benchmark>();
     }
 }