This PR is a draft of adding the ability for the crank controller to produce a comparison report for BenchmarkDotNet results that can be used with the `crank-pr` tool (see #399).

At the moment this draft uses some code taken from [BenchmarkDotNet](https://github.com/dotnet/BenchmarkDotNet) and [Perfolizer](https://github.com/AndreyAkinshin/perfolizer) and trimmed down to just the parts that are needed in order to handle unit conversions and rounding to improve the UX compared to rendering everything as nanoseconds and bytes using the raw values from the JSON output. It would be preferable to just reference BenchmarkDotNet itself, but I've left out adding the additional dependency to do that for now.

Example JSON files I've been using to test are included below, which give the following comparison outputs.

Note that in most of the examples memory allocations are missing due to dotnet/BenchmarkDotNet#1919. The ones in example 1 are fake values I manually added to the JSON files I used for testing.

## Example 1

```md

| ArrayBenchmarks.Empty | base   | pr     |   |
| --------------------- | ------ | ------ | - |
| Mean (ns)             | 0.0110 | 0.0000 |   |
| Error (ns)            | 0.0110 | 0.0000 |   |
| StdDev (ns)           | 0.0191 | 0.0000 |   |
| Median (ns)           | 0.0000 | 0.0000 |   |


| EmptyBenchmarks.ArrayEmpty | base   | pr     |     |
| -------------------------- | ------ | ------ | --- |
| Mean (ns)                  | 0.0000 | 0.0013 | +8% |
| Error (ns)                 | 0.0000 | 0.0013 | +8% |
| StdDev (ns)                | 0.0000 | 0.0022 | +8% |
| Median (ns)                | 0.0000 | 0.0000 |     |


| EnumerableBenchmarks.Empty | base   | pr     |   |
| -------------------------- | ------ | ------ | - |
| Mean (ns)                  | 0.0000 | 0.0000 |   |
| Error (ns)                 | 0.0000 | 0.0000 |   |
| StdDev (ns)                | 0.0000 | 0.0000 |   |
| Median (ns)                | 0.0000 | 0.0000 |   |


| StringBenchmarks.ConstructorArray | base    | pr      |          |
| --------------------------------- | ------- | ------- | -------- |
| Mean (ns)                         | 11.5654 | 10.5742 |   -8.57% |
| Error (ns)                        |  0.1231 |  0.2537 | +106.02% |
| StdDev (ns)                       |  0.2133 |  0.4394 | +106.02% |
| Median (ns)                       | 11.6748 | 10.5523 |   -9.61% |
| Gen 0                             |     513 |     513 |    0.00% |
| Allocated (B)                     |      64 |      32 |  -50.00% |


| EmptyBenchmarks.EnumerableEmpty | base   | pr     |   |
| ------------------------------- | ------ | ------ | - |
| Mean (ns)                       | 0.0000 | 0.0000 |   |
| Error (ns)                      | 0.0000 | 0.0000 |   |
| StdDev (ns)                     | 0.0000 | 0.0000 |   |
| Median (ns)                     | 0.0000 | 0.0000 |   |


| StringBenchmarks.ConstructorSpan | base    | pr      |         |
| -------------------------------- | ------- | ------- | ------- |
| Mean (ns)                        | 10.7638 | 10.2041 |  -5.20% |
| Error (ns)                       |  0.0529 |  0.0700 | +32.30% |
| StdDev (ns)                      |  0.0916 |  0.1212 | +32.30% |
| Median (ns)                      | 10.7170 | 10.2466 |  -4.39% |
| Gen 0                            |     513 |     513 |   0.00% |
| Allocated (KB)                   |    62.5 |   31.25 | -50.00% |


```

<details>

<summary>base.json</summary>

```json
{
  "Benchmarks": [
    {
      "DisplayInfo": "ArrayBenchmarks.Empty: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "ArrayBenchmarks",
      "Method": "Empty",
      "MethodTitle": "Empty",
      "Parameters": "",
      "FullName": "Microbenchmarks.ArrayBenchmarks.Empty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0.03300756216049194
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": -0.024755671620368958,
        "Q1": 0,
        "Median": 0,
        "Mean": 0.01100252072016398,
        "Q3": 0.01650378108024597,
        "UpperFence": 0.04125945270061493,
        "Max": 0.03300756216049194,
        "InterquartileRange": 0.01650378108024597,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0.01100252072016398,
        "Variance": 0.0003631663865929132,
        "StandardDeviation": 0.019056924898653328,
        "Skewness": 0.38490017945975064,
        "Kurtosis": 0.6666666666666669,
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0.01100252072016398,
          "StandardError": 0.01100252072016398,
          "Level": 12,
          "Margin": 0.34766925271463134,
          "Lower": -0.33666673199446734,
          "Upper": 0.35867177343479534
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0.011222571134567264,
          "P80": 0.01980453729629517,
          "P85": 0.02310529351234436,
          "P90": 0.026406049728393555,
          "P95": 0.029706805944442746,
          "P100": 0.03300756216049194
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 134217728
      }
    },
    {
      "DisplayInfo": "EmptyBenchmarks.ArrayEmpty: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "EmptyBenchmarks",
      "Method": "ArrayEmpty",
      "MethodTitle": "ArrayEmpty",
      "Parameters": "",
      "FullName": "Microbenchmarks.EmptyBenchmarks.ArrayEmpty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": 0,
        "Q1": 0,
        "Median": 0,
        "Mean": 0,
        "Q3": 0,
        "UpperFence": 0,
        "Max": 0,
        "InterquartileRange": 0,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0,
        "Variance": 0,
        "StandardDeviation": 0,
        "Skewness": "",
        "Kurtosis": "",
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0,
          "StandardError": 0,
          "Level": 12,
          "Margin": 0,
          "Lower": 0,
          "Upper": 0
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0,
          "P80": 0,
          "P85": 0,
          "P90": 0,
          "P95": 0,
          "P100": 0
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 134217728
      }
    },
    {
      "DisplayInfo": "EnumerableBenchmarks.Empty: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "EnumerableBenchmarks",
      "Method": "Empty",
      "MethodTitle": "Empty",
      "Parameters": "",
      "FullName": "Microbenchmarks.EnumerableBenchmarks.Empty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": 0,
        "Q1": 0,
        "Median": 0,
        "Mean": 0,
        "Q3": 0,
        "UpperFence": 0,
        "Max": 0,
        "InterquartileRange": 0,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0,
        "Variance": 0,
        "StandardDeviation": 0,
        "Skewness": "",
        "Kurtosis": "",
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0,
          "StandardError": 0,
          "Level": 12,
          "Margin": 0,
          "Lower": 0,
          "Upper": 0
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0,
          "P80": 0,
          "P85": 0,
          "P90": 0,
          "P95": 0,
          "P100": 0
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 134217728
      }
    },
    {
      "DisplayInfo": "StringBenchmarks.ConstructorArray: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "StringBenchmarks",
      "Method": "ConstructorArray",
      "MethodTitle": "ConstructorArray",
      "Parameters": "",
      "FullName": "Microbenchmarks.StringBenchmarks.ConstructorArray",
      "Statistics": {
        "OriginalValues": [
          11.674836277961731,
          11.701834201812744,
          11.319676041603088
        ],
        "N": 3,
        "Min": 11.319676041603088,
        "LowerFence": 11.210637539625168,
        "Q1": 11.49725615978241,
        "Median": 11.674836277961731,
        "Mean": 11.565448840459188,
        "Q3": 11.688335239887238,
        "UpperFence": 11.97495386004448,
        "Max": 11.701834201812744,
        "InterquartileRange": 0.19107908010482788,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0.12313329290682898,
        "Variance": 0.04548542346623682,
        "StandardDeviation": 0.21327311941788824,
        "Skewness": -0.3779728968646388,
        "Kurtosis": 0.6666666666666707,
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 11.565448840459188,
          "StandardError": 0.12313329290682898,
          "Level": 12,
          "Margin": 3.8908956427369503,
          "Lower": 7.674553197722238,
          "Upper": 15.456344483196139
        },
        "Percentiles": {
          "P0": 11.319676041603088,
          "P25": 11.49725615978241,
          "P50": 11.674836277961731,
          "P67": 11.684015572071075,
          "P80": 11.691035032272339,
          "P85": 11.69373482465744,
          "P90": 11.696434617042542,
          "P95": 11.699134409427643,
          "P100": 11.701834201812744
        }
      },
      "Memory": {
        "Gen0Collections": 513,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "BytesAllocatedPerOperation": 64,
        "TotalOperations": 67108864
      }
    },
    {
      "DisplayInfo": "EmptyBenchmarks.EnumerableEmpty: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "EmptyBenchmarks",
      "Method": "EnumerableEmpty",
      "MethodTitle": "EnumerableEmpty",
      "Parameters": "",
      "FullName": "Microbenchmarks.EmptyBenchmarks.EnumerableEmpty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": 0,
        "Q1": 0,
        "Median": 0,
        "Mean": 0,
        "Q3": 0,
        "UpperFence": 0,
        "Max": 0,
        "InterquartileRange": 0,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0,
        "Variance": 0,
        "StandardDeviation": 0,
        "Skewness": "",
        "Kurtosis": "",
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0,
          "StandardError": 0,
          "Level": 12,
          "Margin": 0,
          "Lower": 0,
          "Upper": 0
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0,
          "P80": 0,
          "P85": 0,
          "P90": 0,
          "P95": 0,
          "P100": 0
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 268435456
      }
    },
    {
      "DisplayInfo": "StringBenchmarks.ConstructorSpan: Job-DEXEON(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "StringBenchmarks",
      "Method": "ConstructorSpan",
      "MethodTitle": "ConstructorSpan",
      "Parameters": "",
      "FullName": "Microbenchmarks.StringBenchmarks.ConstructorSpan",
      "Statistics": {
        "OriginalValues": [
          10.869413614273071,
          10.717004537582397,
          10.705120861530304
        ],
        "N": 3,
        "Min": 10.705120861530304,
        "LowerFence": 10.587843134999275,
        "Q1": 10.71106269955635,
        "Median": 10.717004537582397,
        "Mean": 10.763846337795258,
        "Q3": 10.793209075927734,
        "UpperFence": 10.91642864048481,
        "Max": 10.869413614273071,
        "InterquartileRange": 0.08214637637138367,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0.052894999247991764,
        "Variance": 0.008393642836335147,
        "StandardDeviation": 0.09161682616383929,
        "Skewness": 0.37762758781537153,
        "Kurtosis": 0.6666666666666666,
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 10.763846337795258,
          "StandardError": 0.052894999247991764,
          "Level": 12,
          "Margin": 1.6714319680568799,
          "Lower": 9.092414369738378,
          "Upper": 12.435278305852137
        },
        "Percentiles": {
          "P0": 10.705120861530304,
          "P25": 10.71106269955635,
          "P50": 10.717004537582397,
          "P67": 10.768823623657227,
          "P80": 10.808449983596802,
          "P85": 10.82369089126587,
          "P90": 10.838931798934937,
          "P95": 10.854172706604004,
          "P100": 10.869413614273071
        }
      },
      "Memory": {
        "Gen0Collections": 513,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "BytesAllocatedPerOperation": 64000,
        "TotalOperations": 67108864
      }
    }
  ]
}
```

</details>

<details>

<summary>pr.json</summary>

```json
{
  "Benchmarks": [
    {
      "DisplayInfo": "ArrayBenchmarks.Empty: Job-UMRKHO(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "ArrayBenchmarks",
      "Method": "Empty",
      "MethodTitle": "Empty",
      "Parameters": "",
      "FullName": "Microbenchmarks.ArrayBenchmarks.Empty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": 0,
        "Q1": 0,
        "Median": 0,
        "Mean": 0,
        "Q3": 0,
        "UpperFence": 0,
        "Max": 0,
        "InterquartileRange": 0,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0,
        "Variance": 0,
        "StandardDeviation": 0,
        "Skewness": "",
        "Kurtosis": "",
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0,
          "StandardError": 0,
          "Level": 12,
          "Margin": 0,
          "Lower": 0,
          "Upper": 0
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0,
          "P80": 0,
          "P85": 0,
          "P90": 0,
          "P95": 0,
          "P100": 0
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 134217728
      }
    },
    {
      "DisplayInfo": "EmptyBenchmarks.ArrayEmpty: Job-UMRKHO(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "EmptyBenchmarks",
      "Method": "ArrayEmpty",
      "MethodTitle": "ArrayEmpty",
      "Parameters": "",
      "FullName": "Microbenchmarks.EmptyBenchmarks.ArrayEmpty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0.0038720667362213135
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": -0.002904050052165985,
        "Q1": 0,
        "Median": 0,
        "Mean": 0.0012906889120737712,
        "Q3": 0.0019360333681106567,
        "UpperFence": 0.004840083420276642,
        "Max": 0.0038720667362213135,
        "InterquartileRange": 0.0019360333681106567,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0.0012906889120737712,
        "Variance": 4.997633603250525E-06,
        "StandardDeviation": 0.002235538772477571,
        "Skewness": 0.38490017945975036,
        "Kurtosis": 0.6666666666666665,
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0.0012906889120737712,
          "StandardError": 0.0012906889120737712,
          "Level": 12,
          "Margin": 0.04078454937379665,
          "Lower": -0.03949386046172288,
          "Upper": 0.04207523828587042
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0.0013165026903152468,
          "P80": 0.0023232400417327885,
          "P85": 0.0027104467153549194,
          "P90": 0.0030976533889770508,
          "P95": 0.0034848600625991817,
          "P100": 0.0038720667362213135
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 134217728
      }
    },
    {
      "DisplayInfo": "EnumerableBenchmarks.Empty: Job-UMRKHO(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "EnumerableBenchmarks",
      "Method": "Empty",
      "MethodTitle": "Empty",
      "Parameters": "",
      "FullName": "Microbenchmarks.EnumerableBenchmarks.Empty",
      "Statistics": {
        "OriginalValues": [
          0,
          0,
          0
        ],
        "N": 3,
        "Min": 0,
        "LowerFence": 0,
        "Q1": 0,
        "Median": 0,
        "Mean": 0,
        "Q3": 0,
        "UpperFence": 0,
        "Max": 0,
        "InterquartileRange": 0,
        "LowerOutliers": [],
        "UpperOutliers": [],
        "AllOutliers": [],
        "StandardError": 0,
        "Variance": 0,
        "StandardDeviation": 0,
        "Skewness": "",
        "Kurtosis": "",
        "ConfidenceInterval": {
          "N": 3,
          "Mean": 0,
          "StandardError": 0,
          "Level": 12,
          "Margin": 0,
          "Lower": 0,
          "Upper": 0
        },
        "Percentiles": {
          "P0": 0,
          "P25": 0,
          "P50": 0,
          "P67": 0,
          "P80": 0,
          "P85": 0,
          "P90": 0,
          "P95": 0,
          "P100": 0
        }
      },
      "Memory": {
        "Gen0Collections": 0,
        "Gen1Collections": 0,
        "Gen2Collections": 0,
        "TotalOperations": 268435456
      }
    },
    {
      "DisplayInfo": "StringBenchmarks.ConstructorArray: Job-UMRKHO(Toolchain=InProcessEmitToolchain, IterationCount=3, LaunchCount=1, WarmupCount=3)",
      "Namespace": "Microbenchmarks",
      "Type": "StringBenchmarks",
      "Method": "ConstructorArray",
      "MethodTitle": "ConstructorArray",
      "Parameters": "",
      "FullName": "Microbenchmarks.StringBenchmarks.ConstructorArray",
…