@sebastienros 2022-02-07
I liked your sample better, with one benchmark per line and the ratio on the same line, understanding it only works with two runs, so when using more runs there will be more tables.

Another way that would look nicer IMO: https://github.com/sebastienros/jint/pull/821 /cc @lahma