@sebastienros 2022-02-07
Note that I think enclosing the tables in triple ticks looks better than when the table is formatted, because the spaces are respected and it's easier to read (like in your examples).