@lahma 2022-02-08
> Another way that would look nicer IMO: [sebastienros/jint#821](https://github.com/sebastienros/jint/pull/821) /cc @lahma

I've used this home-grown hack https://github.com/lahma/BenchmarkDotNet.ResultDiff to produce results. Basically takes two CSV outputs from runs and runs the diffs and outputs the markdown.