@martincostello 2022-02-08
I've changed the format to be more like my original suggestion https://github.com/dotnet/crank/issues/399#issue-1125215911

For the same input files, the examples now render as below.

## Example 1

```md
| benchmark                         | mean (base) | mean (pr)  | ratio    | allocated (base) | allocated (pr) | ratio |
| --------------------------------- | ----------- | ---------- | -------- | ---------------- | -------------- | ----- |
| ArrayBenchmarks.Empty             |   0.0110 ns |  0.0000 ns |          |              0 B |            0 B |       |
| EmptyBenchmarks.ArrayEmpty        |   0.0000 ns |  0.0013 ns | Infinity |              0 B |            0 B |       |
| EnumerableBenchmarks.Empty        |   0.0000 ns |  0.0000 ns |          |              0 B |            0 B |       |
| StringBenchmarks.ConstructorArray |  11.5654 ns | 10.5742 ns |     0.91 |             64 B |           32 B |  0.50 |
| EmptyBenchmarks.EnumerableEmpty   |   0.0000 ns |  0.0000 ns |          |              0 B |            0 B |       |
| StringBenchmarks.ConstructorSpan  |  10.7638 ns | 10.2041 ns |     0.95 |          62.5 KB |       31.25 KB |  0.50 |
```

## Example 2

```md
| benchmark                  | mean (base) | mean (pr)     | ratio |
| -------------------------- | ----------- | ------------- | ----- |
| Solve2021(input: Y2021-01) | 190.4202 µs |   200.8045 µs |  1.05 |
| Solve2021(input: Y2021-02) | 232.6159 µs |   265.6511 µs |  1.14 |
| Solve2021(input: Y2021-03) | 337.0687 µs |   372.0647 µs |  1.10 |
| Solve2021(input: Y2021-04) | 966.9915 µs | 1,016.2627 µs |  1.05 |
| Solve2021(input: Y2021-05) |  29.3216 ms |    31.8954 ms |  1.09 |
| Solve2021(input: Y2021-06) | 116.0845 µs |   124.3489 µs |  1.07 |
| Solve2021(input: Y2021-07) |  34.2316 ms |    37.6925 ms |  1.10 |
| Solve2021(input: Y2021-08) |   3.2540 ms |     3.4847 ms |  1.07 |
| Solve2021(input: Y2021-09) |   6.4363 ms |     6.6342 ms |  1.03 |
| Solve2021(input: Y2021-10) | 156.4398 µs |   165.1050 µs |  1.06 |
| Solve2021(input: Y2021-11) |   8.2151 ms |     8.7839 ms |  1.07 |
| Solve2021(input: Y2021-12) | 111.7994 ms |   114.8952 ms |  1.03 |
| Solve2021(input: Y2021-13) | 722.1174 µs |   762.6098 µs |  1.06 |
| Solve2021(input: Y2021-14) | 705.5314 µs |   760.3059 µs |  1.08 |
| Solve2021(input: Y2021-15) | 259.4354 ms |   274.2191 ms |  1.06 |
| Solve2021(input: Y2021-16) |  92.0474 µs |    95.7234 µs |  1.04 |
| Solve2021(input: Y2021-17) |  11.5101 ms |    12.2882 ms |  1.07 |
| Solve2021(input: Y2021-18) | 219.1846 ms |   226.8097 ms |  1.03 |
| Solve2021(input: Y2021-19) |    1.6898 s |      1.9367 s |  1.15 |
| Solve2021(input: Y2021-20) | 572.1884 ms |   582.6223 ms |  1.02 |
| Solve2021(input: Y2021-21) | 243.4158 ms |   263.2657 ms |  1.08 |
| Solve2021(input: Y2021-22) | 235.9510 ms |   249.1028 ms |  1.06 |
| Solve2021(input: Y2021-23) | 893.9351 ms |   931.4017 ms |  1.04 |
| Solve2021(input: Y2021-24) |  31.3436 µs |    32.9728 µs |  1.05 |
| Solve2021(input: Y2021-25) |  79.6206 ms |    85.1465 ms |  1.07 |
```