@lahma 2022-02-08
Just a suggestion. Ideally first data column after name would directly tell if interesting. 

Like `Diff` and value like `+1.2%, Mem: -10.2%` - would reduce need for horizontal scrolling when checking results.