@martincostello 2022-02-08
For reference, I'm doing local E2E debugging using the following debug profile:

```json
{
  "profiles": {
    "Microsoft.Crank.PullRequestBot": {
      "commandName": "Project",
      "commandLineArgs": "--benchmarks microbenchmarks --components sandbox --config <path of martincostello/crank-sandbox clone>\\benchmark.yml --profiles local --pull-request 21 --repository https://github.com/martincostello/crank-sandbox",
      "workingDirectory": "<path of martincostello/crank-sandbox clone>"
    }
  }
}
```