@martincostello 2022-02-08
I went with a ratio rather than a percentage or absolute number for consistency with how BDN reports them itself when you have a [benchmark with a baseline](https://github.com/martincostello/crank-sandbox/blob/f9f6c45d5c6ddbb46988cfca1b4dca55e3be6523/src/Microbenchmarks/StringBenchmarks.cs#L12-L16).

For example if you run [this benchmark project](https://github.com/martincostello/crank-sandbox/tree/main/src/Microbenchmarks) with the command below, you'll get something like this with `Alloc Ratio` (which I can rename the column in this PR to match).

```sh
dotnet run --configuration Release -- --filter *Constructor* --memory --job short
```

``` ini

BenchmarkDotNet=v0.13.1.1695-nightly, OS=Windows 11 (10.0.22000.469/21H2)
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.101
  [Host]   : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  ShortRun : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
|           Method |     Mean |    Error |    StdDev | Ratio | RatioSD |  Gen 0 | Allocated | Alloc Ratio |
|----------------- |---------:|---------:|----------:|------:|--------:|-------:|----------:|------------:|
| ConstructorArray | 9.081 ns | 8.902 ns | 0.4879 ns |  1.00 |    0.00 | 0.0076 |      64 B |        1.00 |
|  ConstructorSpan | 8.856 ns | 6.629 ns | 0.3634 ns |  0.98 |    0.07 | 0.0076 |      64 B |        1.00 |