@sebastienros 2022-02-08
I meant (ratio | percentage)  vs. absolute value. I think it's fine to have it, just that seeing the ratio was new to me on this column. Didn't even know you could do that.

I will merge once other PRs are ready, there is this rule that makes me re-build when the base (main) has changed on the PR.