@martincostello 2022-02-08
Ah I see - yeah, I've not seen it often but yeah it's useful at-a-glance.

As this'll need to rebuild anyway, do you want me to push up a commit to rename the second ratio column to match BDN (`ratio` to `alloc ratio`)?