@martincostello 2022-02-08
I noticed I wasn't passing the precision through for the times. I've fixed that now, so it better matches BDN and saves a bit of width.

### Example 1


```md
| benchmark                         | mean (base) | mean (pr) | ratio    | allocated (base) | allocated (pr) | ratio |
| --------------------------------- | ----------- | --------- | -------- | ---------------- | -------------- | ----- |
| ArrayBenchmarks.Empty             |   0.0110 ns | 0.0000 ns |          |              0 B |            0 B |       |
| EmptyBenchmarks.ArrayEmpty        |   0.0000 ns | 0.0013 ns | Infinity |              0 B |            0 B |       |
| EmptyBenchmarks.EnumerableEmpty   |      0.0 ns |    0.0 ns |          |              0 B |            0 B |       |
| EnumerableBenchmarks.Empty        |      0.0 ns |    0.0 ns |          |              0 B |            0 B |       |
| StringBenchmarks.ConstructorArray |    11.57 ns |  10.57 ns |     0.91 |             64 B |           32 B |  0.50 |
| StringBenchmarks.ConstructorSpan  |    10.76 ns |  10.20 ns |     0.95 |          62.5 KB |       31.25 KB |  0.50 |
```

### Example 2

```md
| benchmark                  | mean (base) | mean (pr)  | ratio |
| -------------------------- | ----------- | ---------- | ----- |
| Solve2021(input: Y2021-01) |    190.4 µs |   200.8 µs |  1.05 |
| Solve2021(input: Y2021-02) |    232.6 µs |   265.7 µs |  1.14 |
| Solve2021(input: Y2021-03) |    337.1 µs |   372.1 µs |  1.10 |
| Solve2021(input: Y2021-04) |    967.0 µs | 1,016.3 µs |  1.05 |
| Solve2021(input: Y2021-05) |     29.3 ms |    31.9 ms |  1.09 |
| Solve2021(input: Y2021-06) |    116.1 µs |   124.3 µs |  1.07 |
| Solve2021(input: Y2021-07) |     34.2 ms |    37.7 ms |  1.10 |
| Solve2021(input: Y2021-08) |      3.3 ms |     3.5 ms |  1.07 |
| Solve2021(input: Y2021-09) |      6.4 ms |     6.6 ms |  1.03 |
| Solve2021(input: Y2021-10) |    156.4 µs |   165.1 µs |  1.06 |
| Solve2021(input: Y2021-11) |      8.2 ms |     8.8 ms |  1.07 |
| Solve2021(input: Y2021-12) |    111.8 ms |   114.9 ms |  1.03 |
| Solve2021(input: Y2021-13) |    722.1 µs |   762.6 µs |  1.06 |
| Solve2021(input: Y2021-14) |    705.5 µs |   760.3 µs |  1.08 |
| Solve2021(input: Y2021-15) |    259.4 ms |   274.2 ms |  1.06 |
| Solve2021(input: Y2021-16) |     92.0 µs |    95.7 µs |  1.04 |
| Solve2021(input: Y2021-17) |     11.5 ms |    12.3 ms |  1.07 |
| Solve2021(input: Y2021-18) |    219.2 ms |   226.8 ms |  1.03 |
| Solve2021(input: Y2021-19) |       1.7 s |      1.9 s |  1.15 |
| Solve2021(input: Y2021-20) |    572.2 ms |   582.6 ms |  1.02 |
| Solve2021(input: Y2021-21) |    243.4 ms |   263.3 ms |  1.08 |
| Solve2021(input: Y2021-22) |    236.0 ms |   249.1 ms |  1.06 |
| Solve2021(input: Y2021-23) |    893.9 ms |   931.4 ms |  1.04 |
| Solve2021(input: Y2021-24) |     31.3 µs |    33.0 µs |  1.05 |
| Solve2021(input: Y2021-25) |     79.6 ms |    85.1 ms |  1.07 |
```