@sebastienros 2022-02-09
This is how it looks like for EF, I can see the allocated column but not in the comparison:


```
[04:23:21.799] Deleting job 'application' ...


BenchmarkDotNet=v0.13.1, OS=Windows 10.0.14393.4825 (1607/AnniversaryUpdate/Redstone1)
Intel Xeon Gold 5120 CPU 2.20GHz, 1 CPU, 28 logical and 14 physical cores
Frequency=2143402 Hz, Resolution=466.5480 ns, Timer=TSC
.NET Core SDK=        2.2.105 [C:\Program Files\dotnet\sdk]
  [Host] : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

Toolchain=InProcessEmitToolchain  InvocationCount=1  IterationCount=3
LaunchCount=1  UnrollFactor=1  WarmupCount=3

      Method | AutoDetectChanges |     Mean |     Error |  StdDev |  Op/s |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
------------ |------------------ |---------:|----------:|--------:|------:|----------:|----------:|----------:|----------:|
         Add |             False | 164.2 ms | 113.78 ms | 6.24 ms | 6.091 | 4000.0000 | 2000.0000 |         - |     48 MB |
    AddRange |             False | 155.4 ms |   6.20 ms | 0.34 ms | 6.436 | 4000.0000 | 1000.0000 |         - |     48 MB |
      Attach |             False | 120.9 ms |  27.66 ms | 1.52 ms | 8.273 | 4000.0000 | 2000.0000 | 1000.0000 |     40 MB |
 AttachRange |             False | 119.9 ms |  26.20 ms | 1.44 ms | 8.341 | 3000.0000 | 1000.0000 |         - |     39 MB |
         Add |              True | 175.2 ms |  62.98 ms | 3.45 ms | 5.707 | 4000.0000 | 1000.0000 |         - |     48 MB |
    AddRange |              True | 161.0 ms |  59.29 ms | 3.25 ms | 6.211 | 5000.0000 | 2000.0000 | 1000.0000 |     48 MB |
      Attach |              True | 121.0 ms |  26.79 ms | 1.47 ms | 8.265 | 4000.0000 | 2000.0000 | 1000.0000 |     40 MB |
 AttachRange |              True | 119.0 ms |  27.19 ms | 1.49 ms | 8.406 | 4000.0000 | 2000.0000 | 1000.0000 |     39 MB |


Results saved in 'C:\Users\sebros.REDMOND\AppData\Local\Temp\adddatavariations.pr.json'

crank compare C:\Users\sebros.REDMOND\AppData\Local\Temp\adddatavariations.base.json C:\Users\sebros.REDMOND\AppData\Local\Temp\adddatavariations.pr.json


| benchmark                             | mean (adddatavariations.base) | mean (adddatavariations.pr) | ratio |
| ------------------------------------- | ----------------------------- | --------------------------- | ----- |
| Add(AutoDetectChanges: False)         |                      153.7 ms |                    164.2 ms |  1.07 |
| Add(AutoDetectChanges: True)          |                      152.5 ms |                    175.2 ms |  1.15 |
| AddRange(AutoDetectChanges: False)    |                      152.6 ms |                    155.4 ms |  1.02 |
| AddRange(AutoDetectChanges: True)     |                      153.3 ms |                    161.0 ms |  1.05 |
| Attach(AutoDetectChanges: False)      |                      116.6 ms |                    120.9 ms |  1.04 |
| Attach(AutoDetectChanges: True)       |                      116.8 ms |                    121.0 ms |  1.04 |
| AttachRange(AutoDetectChanges: False) |                      115.6 ms |                    119.9 ms |  1.04 |
| AttachRange(AutoDetectChanges: True)  |                      115.6 ms |                    119.0 ms |  1.03 |

```